fixed some compilation problems with the school norm flags and updated notes

+++ libft/ft_strdup.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: echoukri <echoukri@student.42.fr>          +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 12:49:22 by echoukri          #+#    #+#             */
-/*   Updated: 2022/10/17 17:16:43 by echoukri         ###   ########.fr       */
+/*   Updated: 2022/10/19 01:08:04 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(char *src)
 {
 	char	*ptr;
-	int		i;
+	unsigned long		i;
 
 	ptr = malloc(sizeof(char) * (ft_strlen(src) + 1));
 	if (!ptr)
